BonusPoint UserWallet

+++ oas-server/oas/src/main/java/com/cascv/oas/server/blockchain/model/UserWallet.java
@@ -1,7 +1,7 @@
 package com.cascv.oas.server.blockchain.model;
 
 import java.io.Serializable;
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -10,5 +10,5 @@ public class UserWallet implements Serializable{
   private static final long serialVersionUID = 1L;
   @Getter @Setter private Integer id;
   @Getter @Setter private Integer useId;
-  @Getter @Setter private BigInteger balance;
+  @Getter @Setter private BigDecimal balance;
 }
