+++ oas-server/oas/src/main/java/com/cascv/oas/server/blockchain/model/UserWalletDetail.java
@@ -1,6 +1,6 @@
 package com.cascv.oas.server.blockchain.model;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -8,7 +8,7 @@ import lombok.Setter;
 public class UserWalletDetail {
   @Getter @Setter private Integer id;
   @Getter @Setter private Integer userWalletId;
-  @Getter @Setter private BigInteger value;
+  @Getter @Setter private BigDecimal value;
   @Getter @Setter private String datetime;
   @Getter @Setter private String name;
 }
